moved adapter-test into tests directory. Seems more logic
