+++ test/adapter-tests.js
